allow toggling attract screens
